JVM: Add linenumber for closing brace in suspend functions
In order to support step-over state-machine builder
moves the next linenumber instruction closer to a suspend
call. However, if the call is the last statement inside
an `if` branch, next linenumber is outside the branch,
and it will be pulled inside the branch. So, from
debugger's point of view the first statement after the if
will appear as if it is inside the if.
Adding linenumber for closing brace mitigates the issue,
but does not solve it completely, since one can use
single-line statement and omit the braces.

 #KT-48311 Fixed